requirements.txt
Flask
transformers
torch
Pillow
pygame
gTTS
googletrans==3.1.0a0
